+++ inst/tinytest/test_intervals.R
@@ -9,6 +9,14 @@ upper <- c(750, 825, 1250, 1275, 1325, 700, 1300, 1325,
 x <- intervals(start = lower, end = upper, calendar = CE())
 expect_identical(length(x), 15L)
 
+expect_error(names(x) <- LETTERS[1:3])
+names(x) <- LETTERS[1:15]
+expect_identical(names(x), LETTERS[1:15])
+expect_identical(labels(x), LETTERS[1:15])
+names(x) <- NULL
+expect_identical(names(x), paste0("I", 1:15))
+expect_identical(labels(x), paste0("I", 1:15))
+
 expect_error(intervals(start = upper, end = lower, calendar = CE()), "is later than")
 
 lower_rd <- fixed(lower, calendar = CE())
@@ -27,6 +35,10 @@ expect_identical(span(x, calendar = CE()), span(x, calendar = BP()))
 # Overlap ======================================================================
 expect_identical(overlap(x, calendar = CE()), overlap(x, calendar = BP()))
 
+names(x) <- c("I1", "I1", "I3", "I4", "I5", "I6", "I7", "I8", "I9", "I10", "I11", "I12", "I13", "I14", "I15")
+over <- overlap(x, calendar = CE(), aggregate = TRUE)
+expect_identical(over[c(1, 5), 1], c(I1 = 352, I6 = 77))
+
 # Inf boundaries ===============================================================
 y <- intervals(start = c(50, -Inf, -Inf), end = c(Inf, 50, Inf), calendar = CE())
 expect_identical(span(y, calendar = CE()), c(Inf, Inf, Inf))
@@ -36,25 +48,20 @@ expect_equivalent(
 )
 
 # Plot =========================================================================
-if (at_home()) {
-  using("tinysnapshot")
-  options(tinysnapshot_device = "svglite")
-  options(tinysnapshot_height = 7) # inches
-  options(tinysnapshot_width = 7)
-  options(tinysnapshot_tol = 200) # pixels
-  options(tinysnapshot_os = "Linux")
+using("tinysnapshot")
+source("helpers.R")
 
-  plot_interval_rd <- function() plot(x, calendar = NULL)
-  expect_snapshot_plot(plot_interval_rd, "plot_interval_rd")
+x <- intervals(start = lower, end = upper, calendar = CE())
 
-  plot_interval_CE <- function() plot(x, calendar = CE())
-  expect_snapshot_plot(plot_interval_CE, "plot_interval_CE")
+plot_interval_rd <- function() plot(x, calendar = NULL)
+expect_snapshot_plot(plot_interval_rd, "plot_interval_rd")
 
-  plot_interval_Inf <- function() plot(y, calendar = CE())
-  expect_snapshot_plot(plot_interval_Inf, "plot_interval_Inf")
+plot_interval_CE <- function() plot(x, calendar = CE())
+expect_snapshot_plot(plot_interval_CE, "plot_interval_CE")
 
-  grp <- c("A", "A", "B", "B", "B", "A", "D", "D", "D", "C", "C",
-           "C", "C", "B", "B")
-  plot_interval_groups <- function() plot(x, calendar = CE(), groups = grp)
-  expect_snapshot_plot(plot_interval_CE, "plot_interval_groups")
-}
+plot_interval_Inf <- function() plot(y, calendar = CE())
+expect_snapshot_plot(plot_interval_Inf, "plot_interval_Inf")
+
+grp <- c("A", "A", "B", "B", "B", "A", "D", "D", "D", "C", "C", "C", "C", "B", "B")
+plot_interval_groups <- function() plot(x, calendar = CE(), groups = grp)
+expect_snapshot_plot(plot_interval_CE, "plot_interval_groups")